requirements.txt
asgiref==3.8.1
attrs==25.3.0
backports.zoneinfo==0.2.1
Django==4.2.23
djangorestframework==3.15.2
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.7.1
exceptiongroup==1.3.0
importlib_resources==6.4.5
inflection==0.5.1
iniconfig==2.1.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
packaging==25.0
pkgutil_resolve_name==1.3.10
pluggy==1.5.0
pytest==8.3.5
pytest-django==4.11.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.1
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.13.2
uritemplate==4.1.1
zipp==3.20.2
